requirements.txt
# Core ML
scikit-learn==1.4.2
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.12.2

# Data
pandas==2.2.1
numpy==1.26.4
scipy==1.13.0

# Experiment tracking & pipeline
mlflow==2.12.1
dvc==3.49.0
dvc-gdrive==3.0.1

# Data validation & drift
evidently==0.4.22
great-expectations==0.18.15

# Explainability
shap==0.45.0

# Serving
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
joblib==1.4.0
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.21.0

# Testing & quality
pytest==8.1.1
pytest-cov==5.0.0
ruff==0.4.1
pre-commit==3.7.0

# Notebooks
jupyter==1.0.0
ipykernel==6.29.4
